+++ lib/common/versionUtils.js
@@ -211,10 +211,8 @@ function getSpecVersion({ type, data, specificationVersion }) {
  */
 function getConcreteSchemaUtils({ type, data, specificationVersion }) {
   const specVersion = getSpecVersion({ type, data, specificationVersion });
-  if (!specVersion) {
-    return;
-  }
   let concreteUtils = {};
+
   if (specVersion === DEFAULT_SPEC_VERSION) {
     concreteUtils = require('../30XUtils/schemaUtils30X');
   }
@@ -246,11 +244,16 @@ function filterOptionsByVersion(options, version) {
  * @returns {boolean} True if the current spec is using swagger 2.0 spec
  */
 function isSwagger(version) {
-  let isSwagger = false;
-  if (version === VERSION_20.version) {
-    isSwagger = true;
-  }
-  return isSwagger;
+  return compareVersion(version, VERSION_20.version);
+}
+
+/**
+ * Calculates if thew current input is using OAS 3.1 spec
+ * @param {string} version The current spec version
+ * @returns {boolean} True if the current spec is using OAS 3.1 spec
+ */
+function isOAS31(version) {
+  return compareVersion(version, VERSION_31.version);
 }
 
 module.exports = {
@@ -261,5 +264,6 @@ module.exports = {
   compareVersion,
   getVersionRegexBySpecificationVersion,
   SWAGGER_VERSION,
-  VERSION_3_1
+  VERSION_3_1,
+  isOAS31
 };
