+++ lib/jsonPointer.js
@@ -9,8 +9,7 @@ const { getKeyInComponents20 } = require('./swaggerUtils/componentParentMatcher'
   escapedTildeString = '~0',
   { getKeyInComponents30 } = require('./30XUtils/componentsParentMatcher'),
   { getKeyInComponents31 } = require('./31XUtils/componentsParentMatcher'),
-  { SWAGGER_VERSION } = require('./common/versionUtils'),
-  { VERSION_3_1 } = require('./common/versionUtils');
+  { isOAS31, isSwagger } = require('./common/versionUtils');
 
 
 /**
@@ -53,15 +52,13 @@ function jsonPointerDecodeAndReplace(filePathName) {
 * @returns {Array} - the calculated keys in an array representing each nesting property name
 */
 function getKeyInComponents(traceFromParent, filePathName, localPath, version) {
-  const localPart = localPath ? `${localPointer}${localPath}` : '',
-    is20 = version === SWAGGER_VERSION,
-    is31 = version === VERSION_3_1;
+  const localPart = localPath ? `${localPointer}${localPath}` : '';
   let result;
 
-  if (is20) {
+  if (isSwagger(version)) {
     result = getKeyInComponents20(traceFromParent, filePathName, localPart, jsonPointerDecodeAndReplace);
   }
-  else if (is31) {
+  else if (isOAS31(version)) {
     result = getKeyInComponents31(traceFromParent, filePathName, localPart, jsonPointerDecodeAndReplace);
   }
   else {
@@ -78,7 +75,7 @@ function getKeyInComponents(traceFromParent, filePathName, localPath, version) {
 * @returns {string} - the concatenated json pointer
 */
 function concatJsonPointer(traceFromParent, targetInRoot) {
-  const traceFromParentAsString = traceFromParent.join('/');
+  const traceFromParentAsString = traceFromParent.join(jsonPointerLevelSeparator);
   return localPointer + targetInRoot + jsonPointerLevelSeparator + traceFromParentAsString;
 }
 
@@ -91,7 +88,7 @@ function concatJsonPointer(traceFromParent, targetInRoot) {
 * @returns {string} - the concatenated json pointer
 */
 function getJsonPointerRelationToRoot(refValue, traceFromKey, version) {
-  let targetInRoot = version === SWAGGER_VERSION ? '' : '/components';
+  let targetInRoot = isSwagger(version) ? '' : '/components';
   if (refValue.startsWith(localPointer)) {
     return refValue;
   }
