+++ lib/relatedFiles.js
@@ -1,7 +1,7 @@
 const parse = require('./parse.js'),
+  traverseUtility = require('traverse'),
   BROWSER = 'browser',
   { DFS } = require('./dfs'),
-  { getReferences } = require('./relatedEntities'),
   { isExtRef, removeLocalReferenceFromPath } = require('./jsonPointer');
 let path = require('path'),
   pathBrowserify = require('path-browserify');
@@ -23,6 +23,9 @@ function comparePaths(path1, path2) {
    * @returns {object} - Detect root files result object
    */
 function calculatePath(parentFileName, referencePath) {
+  if (path.isAbsolute(referencePath)) {
+    return referencePath;
+  }
   let currentDirName = path.dirname(parentFileName),
     refDirName = path.join(currentDirName, referencePath);
   return refDirName;
@@ -59,6 +62,44 @@ function findNodeFromPath(referencePath, allData) {
   return allData.find((node) => {
     return comparePaths(node.fileName, referencePath);
   });
+}
+
+/**
+   * verifies if the path has been added to the result
+   * @param {string} path - path to find
+   * @param {Array} referencesInNode - Array with the already added paths
+   * @returns {boolean} - wheter a node with the same path has been added
+   */
+function added(path, referencesInNode) {
+  return referencesInNode.find((reference) => { return reference.path === path; }) !== undefined;
+}
+
+/**
+   * Gets all the $refs from an object
+   * @param {object} currentNode - current node in process
+   * @param {Function} refTypeResolver - function to resolve the ref according to type (local, external, web etc)
+   * @param {Function} pathSolver - function to resolve the Path
+   * @returns {object} - {path : $ref value}
+   */
+function getReferences (currentNode, refTypeResolver, pathSolver) {
+  let referencesInNode = [];
+  traverseUtility(currentNode).forEach((property) => {
+    if (property) {
+      let hasReferenceTypeKey;
+      hasReferenceTypeKey = Object.keys(property)
+        .find(
+          (key) => {
+            return refTypeResolver(property, key);
+          }
+        );
+      if (hasReferenceTypeKey) {
+        if (!added(property.$ref, referencesInNode)) {
+          referencesInNode.push({ path: pathSolver(property) });
+        }
+      }
+    }
+  });
+  return referencesInNode;
 }
 
 /**
