+++ lib/schemaUtils.js
@@ -6,7 +6,7 @@
 const { ParseError } = require('./common/ParseError.js');
 
 const { formatDataPath, checkIsCorrectType, isKnownType } = require('./common/schemaUtilsCommon.js'),
-  { getConcreteSchemaUtils, SWAGGER_VERSION } = require('./common/versionUtils.js'),
+  { getConcreteSchemaUtils, isSwagger } = require('./common/versionUtils.js'),
   async = require('async'),
   sdk = require('postman-collection'),
   schemaFaker = require('../assets/json-schema-faker.js'),
@@ -4849,7 +4849,7 @@ module.exports = {
   mapBundleOutput(format, parsedRootFiles, version) {
     return (contentAndComponents) => {
       let bundledFile = contentAndComponents.fileContent;
-      if (version === SWAGGER_VERSION) {
+      if (isSwagger(version)) {
         Object.entries(contentAndComponents.components).forEach(([key, value]) => {
           bundledFile[key] = value;
         });
@@ -4912,7 +4912,7 @@ module.exports = {
         let parsedContent = parseFileOrThrow(rootFile.content);
         return { fileName: rootFile.fileName, content: rootFile.content, parsed: parsedContent };
       }).filter((rootWithParsedContent) => {
-        let fileVersion = version === SWAGGER_VERSION ?
+        let fileVersion = isSwagger(version) ?
           rootWithParsedContent.parsed.oasObject.swagger :
           rootWithParsedContent.parsed.oasObject.openapi;
         return compareVersion(version, fileVersion);
@@ -4920,8 +4920,8 @@ module.exports = {
       data = toBundle ?
         this.getBundledFileData(parsedRootFiles, inputData, origin, bundleFormat, version) :
         this.getRelatedFilesData(parsedRootFiles, inputData, origin);
-    return data;
 
+    return data;
   },
 
   /**
